remove HostiePuppet from wechaty_puppet_service source code
