+++ src/wechaty_puppet_service/__init__.py
@@ -1,11 +1,9 @@
 """
 doc
 """
-from .puppet import HostiePuppet
 from .puppet import PuppetService
 
 
 __all__ = [
-    'HostiePuppet',
     'PuppetService'
 ]
